requirements.txt
numpy<2.0
pandas
matplotlib
jupyter
ipykernel
scikit-learn
seaborn
matplotlib
load_dotenv
psycopg2
sqlalchemy
